pyproject.toml
[tool.poetry]
name = "sourander-github-io"
version = "0.1.1"
description = "A portal or front page for all Github Pages sites under sourander.github.io"
authors = ["Jani Sourander <jani.sourander@kamk.fi>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "^9.5.9"
jinja2 = "^3.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
write-index = "sourander_github_io.main:write_index"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
